Various fixes.
DivGeo.py can now be rerun, if the user change the DEVICE environment.

Carre.py now has lns in the check boxes if there is an error and a lns
needs to be run again. Because it is hard to make a simple signal to
update the list of .dg files in baserun, a push button is implemented
so the user can update the list of .dg files in baserun.