Versions are forced into build scripts from Makefile, with the help of sed command. Ie, if the version of OpenBLAS is changed to 0.3.4 from the Makefile, then when running make, the version will be also set as the default version in the biuld script.
Added scripts to build libraries that cannot be installed by a user on
the ITER cluster (requires root privileges):

HDF5
NetCDF-C and NetCDF-Fortran
cURL (for NetCDF DAP support)
OpenMPI
Flex (build-time)
FreeType2
NCL (or NCARG)

OpenBLAS now has the DYNAMIC_ARCH=1 flag when compiling, so that it
compiles all kernels and loads the suitable one on run time.

Force solps-iter source.csh script to always select the configuration
for UNKNOWN machine (forcing gfortran with custom locations for
external packages).

NetCDF and NetCDF-Fortran installs into the same directory. On CentOS
the lib install directory of NetCDF-Fortran is ``lib64``, which is
annoying.